+++ hexarena/Scripts/hexarena/Interfaces/ITargetable.cs
@@ -4,7 +4,7 @@ namespace Interfaces
 {
 	public interface ITargetable
 	{
-		public abstract void TakeDamage(short damage);
+		public abstract void TakeDamage(int damage);
 
 		public abstract void TakeStatusEffect(IStatusEffect effect);
